requirements.txt
streamlit
langchain
faiss-cpu
langchain-community
python-dotenv
pypdf
bs4
pymupdf
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
langchain_groq
